feat: new switch component
